pyproject.toml
[project]
name = "memoboxproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "langchain-openai>=0.3.8",
    "openai>=1.0.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.7.0",
    "ruff>=0.11.2",
    "django-filter>=25.1",
    "drf-yasg>=1.21.10",
]
